Separates the concers of rendering and tiletype/rendering associtations to infant rendering engine and away from ProceduralMapGenerator class.
